requirements.txt
aiosqlite==0.21.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asttokens==3.0.0
attrs==25.1.0
auto-py-to-exe==2.45.0
beautifulsoup4==4.12.3
bottle==0.13.2
bottle-websocket==0.2.9
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cssselect==1.2.0
decorator==5.1.1
dnspython==2.7.0
docstring_parser==0.16
dotenv==0.9.9
Eel==0.18.1
execnet==2.1.1
executing==2.2.0
fake-useragent==2.0.3
feedparser==6.0.11
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
gigachat==0.1.38
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.77.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
ipython==8.32.0
jedi==0.19.2
lxml==5.3.1
lxml_html_clean==0.4.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
parso==0.8.4
pefile==2023.2.7
playwright==1.50.0
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
pyppeteer==0.0.25
pyquery==2.0.1
PySocks==1.7.1
pytest==8.3.5
pytest-dependency==0.6.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
requests-html==0.10.0
rsa==4.9
selenium==4.29.0
sgmllib3k==1.0.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
w3lib==2.3.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
